Mark specs that fail on windows
We fail early on windows when trying to do stuff with the module tool
because `module install` tries to use tar and `module list` outputs
color character sequences that the windows terminal doesn't know how to
deal with.  Until those issues are dealt with the specs around those
areas are known to fail on windows.